completes transition to tableview with more broken out and organized structure
