--- PhotoPoints/View/Points/PointsFooter.swift
@@ -1,13 +1,0 @@
-//
-//  PointsFooterView.swift
-//  PhotoPoints
-//
-//  Created by Clay Suttner on 1/9/21.
-//  Copyright © 2021 Cascadia College. All rights reserved.
-//
-
-import UIKit
-
-class PointsFooterView: UIView {
-    
-}
