--- PhotoPoints/View/Points/PointsHeader.swift
@@ -1,13 +1,0 @@
-//
-//  PointsHeaderView.swift
-//  PhotoPoints
-//
-//  Created by Clay Suttner on 1/9/21.
-//  Copyright © 2021 Cascadia College. All rights reserved.
-//
-
-import UIKit
-
-class PointsHeader: UIView {
-    
-}
